+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 # 0.9.6
-- shortHistoryCount added. A list of instructions is highlighted that were executed just before the current one.
+- spotCount added. A list of instructions is highlighted that were executed just before and just after the current one.
 - Fixed extended and real stack.
 - Fix for coverage display after unit tests stopped.
 - Fix for forward step during reverse debug (isCallOpcode corrected)